changes for ks header
